rename file multigrid_input_parameters.h
