+++ include/exadg/poisson/user_interface/input_parameters.h
@@ -24,7 +24,7 @@
 
 #include <exadg/grid/enum_types.h>
 #include <exadg/poisson/user_interface/enum_types.h>
-#include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
+#include <exadg/solvers_and_preconditioners/multigrid/multigrid_parameters.h>
 #include <exadg/solvers_and_preconditioners/solvers/solver_data.h>
 #include <exadg/utilities/print_functions.h>
