+++ include/exadg/solvers_and_preconditioners/multigrid/smoothers/gmres_smoother.h
@@ -26,7 +26,7 @@
 #include <deal.II/lac/solver_gmres.h>
 
 // ExaDG
-#include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
+#include <exadg/solvers_and_preconditioners/multigrid/multigrid_parameters.h>
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/smoother_base.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/block_jacobi_preconditioner.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/jacobi_preconditioner.h>
