+++ include/exadg/solvers_and_preconditioners/preconditioners/preconditioner_amg.h
@@ -28,7 +28,7 @@
 #include <deal.II/lac/trilinos_precondition.h>
 #include <deal.II/lac/trilinos_sparse_matrix.h>
 
-#include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
+#include <exadg/solvers_and_preconditioners/multigrid/multigrid_parameters.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/preconditioner_base.h>
 #include <exadg/solvers_and_preconditioners/utilities/petsc_operation.h>
 #include <exadg/utilities/print_functions.h>
